requirements.txt
python-telegram-bot==20.6
yfinance==0.2.27
pytz==2025.7
pandas==2.1.1
numpy==1.26.0